% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpath.R
\name{getpath}
\alias{getpath}
\title{Query your study via the PIC-SURE API}
\usage{
getpath(env, key, var, verbose = FALSE)
}
\arguments{
\item{env}{The URL of the environment}

\item{key}{The key or the token to log in your environment}

\item{var}{A vector with the variables of interest. You can put a variable, or a path, as you want. You can also use the {*} key if you want to use a wild card. If an argument corresponds to a node, it will return all the variables below the node}

\item{verbose}{By default, {verbose = FALSE}. Set it to {verbose = TRUE} to get the log informations}

\item{subset}{By default, subset = ALL and gives you back all the patients that have at least one variable of interest. See the examples for more complex subsets. !! Similarly, the "subsetting" part needs to be slightly modified to handle the "find" function from picsure.}

\item{gabe}{if TRUE, will export the JSON query as a query.txt file in you working directory. By default, gabe = FALSE}
}
\value{
Returns a data.frame
}
\description{
For this beta version, it is only possible to query phenotypics data. Soon, the package will be upgraded to be able to query genotypics data from Hail.
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne, Alba Gutierrez
}
