% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.hail.R
\name{queryhail}
\alias{queryhail}
\title{Query on Hail}
\usage{
queryhail(env, key, chromosome, start, stop, verbose = FALSE)
}
\arguments{
\item{env}{The URL of the environment}

\item{key}{The key or the token to log in your environment}
}
\value{
Returns a data.frame
}
\description{
Query on Hail
}
\author{
Gregoire Versmee, Laura Versmee
}
