% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheescake.R
\name{picsure}
\alias{picsure}
\title{TEST}
\usage{
picsure(env, key, var, subset = "ALL", verbose = FALSE)
}
\arguments{
\item{env}{The URL of the environment}

\item{key}{The key to log in your environment
function}

\item{var}{A vector with the variables of interest}

\item{url}{The url.}
}
\description{
TEST
}
\author{
Gregoire Versmee, Laura Versmee, Mikael Dusenne
}
